api/requirements.txt
pyserial==3.5
Flask==2.3.2
Flask-SocketIO==5.3.4
eventlet==0.33.3
